% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_upload}
\alias{meta_upload}
\title{上传元数据}
\usage{
meta_upload(
  conn = tsda::conn_rds("metadata"),
  file_name = "data-raw/data/物料生成列表模板.xlsx",
  sheet_name = "数据字典",
  FDictTableName = "t_dict_multipleMaterial",
  overwrite = TRUE
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{数据表}

\item{sheet_name}{页答名称}

\item{FDictTableName}{数据字典表}

\item{overwrite}{是否重写内码}
}
\value{
返回值
}
\description{
上传元数据
}
\examples{
meta_upload()
}
