% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{metadata_createStat}
\alias{metadata_createStat}
\title{构建视图SQL语句}
\usage{
metadata_createStat(
  conn = tsda::conn_rds("metadata"),
  FFormName = "销售订单",
  FActionDesc = "保存",
  FOwnerName = "CP"
)
}
\arguments{
\item{conn}{连接}

\item{FFormName}{表单名称}

\item{FActionDesc}{操作描述}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
构建视图SQL语句
}
\examples{
metadata_createStat()
}
