% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{metadata_createTable}
\alias{metadata_createTable}
\title{无数据创建表格}
\usage{
metadata_createTable(
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormName}{表单名称}

\item{FActionDesc}{操作名称}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
无数据创建表格
}
\examples{
metadata_createTable()
}
