% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{metadata_createView}
\alias{metadata_createView}
\title{构建视图SQL语句}
\usage{
metadata_createView(
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormName}{表单名称}

\item{FActionDesc}{操作描述}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
构建视图SQL语句
}
\examples{
metadata_createView()
}
