% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{metadata_getDbName}
\alias{metadata_getDbName}
\title{获取数据库名称}
\usage{
metadata_getDbName(
  conn = tsda::conn_rds("willingox"),
  FToken = "970C9E78-F276-4073-95AB-790A43BB38AB"
)
}
\arguments{
\item{conn}{连接}

\item{FToken}{凭据}
}
\value{
返回值
}
\description{
获取数据库名称
}
\examples{
metadata_getDbName()
}
