% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{metadata_objectIsNew}
\alias{metadata_objectIsNew}
\title{检查对象是否存在}
\usage{
metadata_objectIsNew(
  FToken = "970C9E78-F276-4073-95AB-790A43BB38AB",
  objectName = "rds_bd_material"
)
}
\arguments{
\item{FToken}{凭证}

\item{objectName}{对象表}
}
\value{
返回值
}
\description{
检查对象是否存在
}
\examples{
metadata_objectIsNew()
}
