% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_dataAll}
\alias{model_dataAll}
\title{返回所有数据}
\usage{
model_dataAll(
  json_file = "data-raw/material_save_cp.json",
  conn = tsda::conn_rds("metadata"),
  FFormId = "BD_MATERIAL",
  FFormName = "物料",
  FActionName = "Save",
  FActionDesc = "保存",
  FTableName = "rds_bd_material",
  FViewName = "rds_vw_bd_material",
  FAccessToken = "970C9E78-F276-4073-95AB-790A43BB38AB",
  FTableKey = "FNumber",
  FTableNameAlias = "t1",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{json_file}{json file}

\item{conn}{连接}

\item{FFormId}{表单}

\item{FFormName}{表名}

\item{FActionName}{操作}

\item{FActionDesc}{描述}

\item{FTableName}{表名}

\item{FViewName}{视图表}

\item{FAccessToken}{图表}

\item{FTableKey}{键值}

\item{FTableNameAlias}{别名}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
返回所有数据
}
\examples{
model_dataAll()
}
