% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_dataAllPart}
\alias{model_dataAllPart}
\title{返回所有数据}
\usage{
model_dataAllPart(
  json_file = "data-raw/material_save_cp.json",
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{json_file}{json file}

\item{conn}{连接}

\item{FFormName}{表名}

\item{FActionDesc}{描述}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
返回所有数据
}
\examples{
model_dataAll()
}
