% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_delete}
\alias{model_delete}
\title{模型删除数据}
\usage{
model_delete(
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormName}{表单名称}

\item{FActionDesc}{操作名称}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
模型删除数据
}
\examples{
model_delete()
}
