% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_entity.R
\name{model_entity_delete}
\alias{model_entity_delete}
\title{删除模型中的数据}
\usage{
model_entity_delete(
  conn = tsda::conn_rds("metadata"),
  FOwnerName = "CP",
  FFormName = "客户",
  FActionDesc = "保存"
)
}
\arguments{
\item{conn}{连接}

\item{FOwnerName}{所有者名称}

\item{FFormName}{表名}

\item{FActionDesc}{操作名称}
}
\value{
返回值
}
\description{
删除模型中的数据
}
\examples{
model_entity_delete()
}
