% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_entity.R
\name{model_entity_read}
\alias{model_entity_read}
\title{模型实体读取数据}
\usage{
model_entity_read(
  conn = tsda::conn_rds("metadata"),
  file_name = "data-raw/kdc_entityTable.xlsx"
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{文件名}
}
\value{
返回值
}
\description{
模型实体读取数据
}
\examples{
model_entity_read()
}
