% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_entry_valueType}
\alias{model_entry_valueType}
\title{模型返回主值是否列表}
\usage{
model_entry_valueType(model, entityName, mainKey)
}
\arguments{
\item{model}{模型}

\item{entityName}{entity name}

\item{mainKey}{main key}
}
\value{
返回值
}
\description{
模型返回主值是否列表
}
\examples{
model_entry_valueType()
}
