% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_entryList_auxKey}
\alias{model_entryList_auxKey}
\title{获取子键内容}
\usage{
model_entryList_auxKey(model, entityName, mainKey)
}
\arguments{
\item{model}{模型}

\item{entityName}{entity name}

\item{mainKey}{main key}
}
\value{
返回值
}
\description{
获取子键内容
}
\examples{
model_entry_auxKey()
}
