% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_entryList_dataAll}
\alias{model_entryList_dataAll}
\title{读取模型元数据}
\usage{
model_entryList_dataAll(
  json_file = "data-raw/material_save_cp.json",
  conn = tsda::conn_rds("metadata"),
  FFormId = "BD_MATERIAL",
  FFormName = "物料",
  FActionName = "Save",
  FActionDesc = "保存",
  FTableName = "rds_bd_material",
  FViewName = "rds_vw_bd_material",
  FAccessToken = "970C9E78-F276-4073-95AB-790A43BB38AB",
  FTableKey = "FNumber",
  FTableNameAlias = "t1",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{json_file}{json file}

\item{conn}{连接}

\item{FFormId}{表单}

\item{FFormName}{fornaeme}

\item{FActionName}{动作}

\item{FActionDesc}{a}

\item{FTableName}{t}

\item{FViewName}{v}

\item{FAccessToken}{t}

\item{FTableKey}{t}

\item{FTableNameAlias}{a}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
读取模型元数据
}
\examples{
model_entry_dataAll()
}
