% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_entryList_dataType}
\alias{model_entryList_dataType}
\title{模型返回主值类型}
\usage{
model_entryList_dataType(model, entityName, mainKey)
}
\arguments{
\item{model}{模型}

\item{entityName}{entity name}

\item{mainKey}{main key}
}
\value{
返回值
}
\description{
模型返回主值类型
}
\examples{
model_entry_dataType()
}
