% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_entryList_MainKey}
\alias{model_entryList_MainKey}
\title{返回主键}
\usage{
model_entryList_MainKey(
  model,
  entityName = "SubHeadEntity",
  conn = tsda::conn_rds("metadata"),
  FFormId = "BD_MATERIAL",
  FActionName = "Save",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{model}{模型}

\item{entityName}{名称}

\item{conn}{连接}

\item{FFormId}{表单ID}

\item{FActionName}{操作}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
返回主键
}
\examples{
model_entryList_MainKey()
}
