% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_entryListPart}
\alias{model_entryListPart}
\title{读取模型元数据}
\usage{
model_entryListPart(
  json_file = "data-raw/material_save_cp.json",
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{json_file}{json file}

\item{conn}{连接}

\item{FFormName}{fornaeme}

\item{FActionDesc}{a}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
读取模型元数据
}
\examples{
model_entry_dataAll()
}
