% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_getEntityList}
\alias{model_getEntityList}
\title{查询表名信息}
\usage{
model_getEntityList(
  conn = tsda::conn_rds("metadata"),
  FOwnerName = "kingdee",
  Ftype = "head",
  FFormName = "物料",
  FActionDesc = "保存"
)
}
\arguments{
\item{conn}{连接}

\item{FOwnerName}{所有者}

\item{Ftype}{类型}

\item{FFormName}{表名}

\item{FActionDesc}{操作名称}
}
\value{
返回值
}
\description{
查询表名信息
}
\examples{
model_getEntityList()
}
