% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_getEntityName}
\alias{model_getEntityName}
\title{获取模型实体名称}
\usage{
model_getEntityName(
  conn = tsda::conn_rds("metadata"),
  FFormId = "BD_MATERIAL",
  FActionName = "Save",
  FType = "entry",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormId}{表单ID}

\item{FActionName}{操作}

\item{FType}{类型}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
获取模型实体名称
}
\examples{
model_getEntityName()
}
