% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_getEntityName_entry}
\alias{model_getEntityName_entry}
\title{模型返回表体}
\usage{
model_getEntityName_entry(
  conn = tsda::conn_rds("metadata"),
  FFormId = "BD_MATERIAL",
  FActionName = "Save",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormId}{表单ID}

\item{FActionName}{操作}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
模型返回表体
}
\examples{
model_getEntityName_entry()
}
