% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_getFromJson}
\alias{model_getFromJson}
\title{读取json数据}
\usage{
model_getFromJson(json_file = "data-raw/material_save_cp.json")
}
\arguments{
\item{json_file}{文件名}
}
\value{
返回值
}
\description{
读取json数据
}
\examples{
model_getFromJson()

}
