% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_getMaxId}
\alias{model_getMaxId}
\title{获取主数据表最大ID}
\usage{
model_getMaxId(conn = tsda::conn_rds("metadata"))
}
\arguments{
\item{conn}{连接}
}
\value{
返顺值
}
\description{
获取主数据表最大ID
}
\examples{
model_getMaxId()
}
