% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_head_auxKey}
\alias{model_head_auxKey}
\title{获取子键内容}
\usage{
model_head_auxKey(model, mainKey)
}
\arguments{
\item{model}{模型}

\item{head_key}{主题}
}
\value{
返回值
}
\description{
获取子键内容
}
\examples{
model_head_subKey_class()
}
