% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_head_dataType}
\alias{model_head_dataType}
\title{模型返回主值类型}
\usage{
model_head_dataType(model, mainKey)
}
\arguments{
\item{model}{模型}

\item{head_key}{键值}
}
\value{
返回值
}
\description{
模型返回主值类型
}
\examples{
model_head_key_class()
}
