% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_head_MainKey}
\alias{model_head_MainKey}
\title{模型获取主键}
\usage{
model_head_MainKey(
  model,
  conn = tsda::conn_rds("metadata"),
  FFormId = "BD_MATERIAL",
  FActionName = "Save",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{model}{模型}

\item{conn}{连接}

\item{FFormId}{表单ID}

\item{FActionName}{操作}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
模型获取主键
}
\examples{
model_head_MainKey()

}
