% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_head_valueType}
\alias{model_head_valueType}
\title{模型返回主值是否列表}
\usage{
model_head_valueType(model, mainKey)
}
\arguments{
\item{model}{模型}

\item{head_key}{键值}
}
\value{
返回值
}
\description{
模型返回主值是否列表
}
\examples{
model_head_valueType()
}
