% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_headPart}
\alias{model_headPart}
\title{模型获取头部数据}
\usage{
model_headPart(
  json_file = "data-raw/material_save_cp.json",
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{json_file}{json file}

\item{conn}{连接}

\item{FFormName}{表单名称}

\item{FActionDesc}{操作名称}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
模型获取头部数据
}
\examples{
model_headPart()
}
