% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_updateCaption}
\alias{model_updateCaption}
\title{模型更新数据更新标题，用于设置对照字典及必录项}
\usage{
model_updateCaption(
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormName}{表单名称}

\item{FActionDesc}{操作名称}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
模型更新数据更新标题，用于设置对照字典及必录项
}
\examples{
model_updateCaption()
}
