% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_updateDataType}
\alias{model_updateDataType}
\title{更新数据类型}
\usage{
model_updateDataType(
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormName}{表名}

\item{FActionDesc}{操作名称}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
更新数据类型
}
\examples{
model_updateDataType()
}
