% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_updateIsShow}
\alias{model_updateIsShow}
\title{更新模型的可见性}
\usage{
model_updateIsShow(
  conn = tsda::conn_rds("metadata"),
  FFormName = "物料",
  FActionDesc = "保存",
  FOwnerName = "kingdee"
)
}
\arguments{
\item{conn}{连接}

\item{FFormName}{表名}

\item{FActionDesc}{操作名}

\item{FOwnerName}{所有者}
}
\value{
返回值
}
\description{
更新模型的可见性
}
\examples{
model_updateIsShow()
}
